Graph View: Enabled vertical cursor when using time on x-axis
 * AP2DataPlot2D - adapted several methods for handling double values as time index and enabled cursor plotting in time axis mode
 * AP2DataPlot2DModel - added method to get a table index for a timestamp and added required data structure
 * AP2DataPlotTread - adapted to new datamodel interface

+++ src/ui/AP2DataPlot2D.h
@@ -160,7 +160,7 @@ private:
     QSortFilterProxyModel *m_tableFilterProxyModel;
     QList<QString> m_tableFilterList;
     int getStatusTextPos();
-    void plotTextArrow(int index, const QString& text, const QString& graph, const QColor& color, QCheckBox *checkBox = NULL);
+    void plotTextArrow(double index, const QString& text, const QString& graph, const QColor& color, QCheckBox *checkBox = NULL);
 
     /**
      * @brief This method hides or shows the text arrows of type
@@ -206,7 +206,7 @@ private:
     /**
      * @brief plotCurrentTime updates the current time red line position
      */
-    void plotCurrentIndex(int index);
+    void plotCurrentIndex(double index);
 
 
 private:
